Handle configuration option for file transformation
It happens that airbrake is not grouping errors (although they are the same errors) because files in backtrace are different. For example they could have different hashs in file endings.  With help of `fileTransformation` object the backtrace file can be transformed (e.g hashs removed) according to a given regular expression.